majit-translate: extract opcode-arm wrapper builders into front::opcode_wrapper
Move the three synthesized-wrapper graph builders (tail-call,
wrapped-unit-variant, raise-stub) and their two shape structs
(WrappedUnitVariantReturn, RaiseStubReturn) out of parse.rs into a new
syn-free front::opcode_wrapper module. Each builder now takes the
dispatcher prologue as an explicit `&[(String, ValueType)]` instead of a
syn::Signature; parse.rs computes that prologue via the new
dispatcher_params_from_sig and delegates. push_dispatcher_inputargs is
folded into the shared builders.

The syn dispatch extractor (parse::extract_match_arms) and the
forthcoming MIR dispatch extractor share one source of truth for the
wrapper graph shapes, so they cannot drift. No behavior change: the syn
synthesizers produce identical graphs.
